refactor: restructure main entry point and refactor device retrieval logic
