+++ src/autodialer/__init__.py
@@ -1,10 +1,8 @@
 __version__ = "0.4.0"
 __author__ = "Byteflow"
 
-from .get_devices import get_devices
-from .reconnection import reconnection
+from .main import main
 
 __all__ = [
-    "get_devices",
-    "reconnection",
+    "main",
 ]
