+++ src/autodialer/get_devices.py
@@ -1,10 +1,6 @@
-import argparse
-import logging
 import sys
 
-from autodialer.routers import get_router
-
-logger = logging.getLogger(__name__)
+from autodialer.routers import RouterAPI
 
 
 def print_devices_table(devices: list) -> None:
@@ -60,30 +56,7 @@ def print_devices_table(devices: list) -> None:
             )
 
 
-def get_devices() -> None:
-    logging.basicConfig(level=logging.INFO, format="%(message)s")
-    parser = argparse.ArgumentParser(
-        description="Get connected devices from the router."
-    )
-    parser.add_argument(
-        "-e",
-        "--env",
-        action="append",
-        help="Set environment variables (e.g., -e PANEL_PASSWORD=secret)",
-    )
-    args = parser.parse_args()
-    if args.env:
-        from autodialer.config import parse_and_save_env_flags
+def get_devices(router: RouterAPI) -> None:
 
-        parse_and_save_env_flags(args.env)
-
-    router = get_router()
-    if router is None:
-        logger.error("Unsupported or undetected router vendor.")
-        sys.exit(1)
     devices = router.get_connected_devices()
     print_devices_table(devices)
-
-
-if __name__ == "__main__":
-    get_devices()
