+++ src/autodialer/network/__init__.py
@@ -1,12 +1,14 @@
 from .check_isp import check_isp_with_retries
+from .get_asn import is_target_asn, normalize_asn, validate_asn
 from .get_ip_address import get_ip_address
-from .is_target_asn import is_target_asn, normalize_asn
 from .wait_internet_recovery import try_connect
 
 __all__ = [
     "check_isp_with_retries",
     "get_ip_address",
-    "is_target_asn",
+    "get_asn",
     "normalize_asn",
+    "validate_asn",
+    "is_target_asn",
     "try_connect",
 ]
